users/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
passlib==1.7.4
bcrypt==4.0.1
tortoise-orm[asyncmy]

# for test purposes
flake8==6.1.0
ipdb==0.13.13
ipython==8.17.2
